iris/requirements.txt
cryptography
PyJWT
python-dotenv
requests>=2.20.0
Werkzeug
PyPDF2
python-docx
pandas
python-pptx
openai
fastapi
uvicorn
starlette
python-multipart
docx
jinja2
tiktoken